update function: 1-create file

+++ 0x15-file_io/1-create_file.c
@@ -12,7 +12,7 @@ int create_file(const char *filename, char *text_content)
 {
 
 int fd;
-ssize_t nr;
+int nr;
 
 nr = 0;
